Move download_so to build.py
