+++ MANIFEST.in
@@ -1,4 +1,4 @@
-recursive-include tests/*
+recursive-include tests *
 
 include ffi/*
 include include/curl/*
